third_party/blink/web_tests/animations/prefixed/animation-events-prefixed-04.html

<!DOCTYPE html>
<title>Tests that custom events with prefixed animations names are correctly dispatched.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
var webkitAnimationStartReceived = false;
var webkitAnimationIterationReceived = false;

document.addEventListener('animationstart', () => {
  assert_unreached("animationstart event listener should not have been called");
});

document.addEventListener('webkitAnimationStart', () => {
  webkitAnimationStartReceived = true;
});

document.addEventListener('animationiteration', () => {
  assert_unreached("animationiteration event listener should not have been called");
});

document.addEventListener('webkitAnimationIteration', () => {
  webkitAnimationIterationReceived = true;
});

document.addEventListener('animationend', () => {
    assert_unreached("animationend event listener should not have been called");
});

async_test(t => {
  document.addEventListener('webkitAnimationEnd', () => {
    t.step(() => {
      assert_true(webkitAnimationIterationReceived && webkitAnimationIterationReceived);
      t.done();
    });
  });

  var custom = document.createEvent('CustomEvent');
  custom.initCustomEvent('webkitAnimationStart', true, true, null);
  document.dispatchEvent(custom);
  custom = document.createEvent('CustomEvent');
  custom.initCustomEvent('webkitAnimationIteration', true, true, null);
  document.dispatchEvent(custom);
  custom = document.createEvent('CustomEvent');
  custom.initCustomEvent('webkitAnimationEnd', true, true, null);
  document.dispatchEvent(custom);
}, "Tests that custom events with prefixed animations names are correctly dispatched");
</script>
